Got eslint working, updated/reformatted files.
To get eslint working I had to add the eslintrc, and juggle some packages.  I've added node and jest to the eslintrc so that it won't complain when we use globally available stuff like it() and expect() in tests.  I've also turned off the jsx-filename-extension linting rule.  I had some trouble getting it to work with .jsx files, so we can stick with .js, at least for now.